Make recursive parsers lazy
Recursive parsers can't be written in an applicative style
as they cause infinite loops. This commit adds a ParserType
protocol which allows different parsers to be defined with
different structures all of which conform to the ParserType
protocol.

Specificallys many and many1 now return ManyParser and Many1Parser
instances.

All of the applicative, functor, and monad funtions and operators
were updated to work with any ParserType not just Parser.

A type erasure struct AnyParser was also created.